github-stats/pyproject.toml
[tool.poetry]
name = "github-stats"
version = "0.1.0"
description = ""
authors = ["Oskar Korczak <oskar.korczak@protonmail.com>"]
readme = "README.md"
packages = [{include = "github_stats"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pygithub = "^2.2.0"
pandas = "^2.2.0"
seaborn = "^0.13.2"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
